Split footer into separate component
